auton broken

+++ src/org/usfirst/frc/team3597/robot/Robot.java
@@ -26,7 +26,7 @@ public class Robot extends IterativeRobot {
 		Robot = new DifferentialDrive(new Spark(LEFTMOTOR), new Spark(RIGHTMOTOR));
 		motor3 = new Talon(9);
 		Controller = new Joystick(1); //Setting up controller to correct port (1)
-		speedMultiplier = (float) 0.8;
+		speedMultiplier = 0.8f;
 	}
 
     long autoStart = 0;
